repotofiles4/requirements.txt


```
python-dotenv
aiohttp
pydantic
langchain
langchain_pinecone
pinecone-client
llama_index
octoai
```